refactor(onboard): finish the code-health split
The first refactor cleared live.go (4.87) and oracle.go (6.79) off the
CodeScene gate. Two files it created stayed under the bar:
live_evidence.go at 6.74 and live_test.go at 6.93.

live_evidence.go still held two concerns that are not evidence: the JSONL
readers, and two filesystem helpers. Those move to live_transcript.go and
live_fileio.go. CaptureEvidence's composer re-read is extracted to
captureEnvironment — it re-reads at evidence time on purpose, because the
environment stored beside a recording has to be the one on screen when the
evidence was taken.

live_test.go grew as the live runs found conditions to cover. It is now four
files matching the source split: composer (including the trust prompt and the
two retries), evidence, archive, and process.

Pure moves and one extraction. Every assertion is unchanged and the package
passes with -race.

Refs #1887
