+++ tools/onboarding-factory/internal/desktopdriver/live_fileio.go
@@ -1,0 +1,30 @@
+package desktopdriver
+
+// Small filesystem helpers shared by the evidence writers.
+
+import (
+	"encoding/json"
+	"fmt"
+	"os"
+)
+
+func writeJSON(path string, value any) error {
+	data, err := json.MarshalIndent(value, "", "  ")
+	if err != nil {
+		return err
+	}
+	data = append(data, '\n')
+	return os.WriteFile(path, data, 0o600)
+}
+
+func copyRegularFile(source, destination string) error {
+	info, err := os.Lstat(source)
+	if err != nil || !info.Mode().IsRegular() {
+		return fmt.Errorf("evidence source is not a regular file: %q", source)
+	}
+	data, err := os.ReadFile(source)
+	if err != nil {
+		return err
+	}
+	return os.WriteFile(destination, data, 0o600)
+}
